requirements.txt
python-dotenv
polygon-api-client
pandas
numpy
pyarrow
scikit-learn
xgboost
joblib
pyyaml
boto3

